Changed Everything. Pretty Much Working.
Not even worth trying to describe changes but now it finds 3 corners and
the puck (using contour detection, not circle detection), does a
perspective shift and draws the prediction, then does a serial output of
that angle to the Arduino.
It serial outputs a 255 if it is unable to find all 4 circles and the
Arduino decides how to pretend it is working.